# services/forge-knowledge/tests/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisBuildRequest": {
        "additionalProperties": false,
        "properties": {
          "concurrency": {
            "default": 1,
            "title": "Concurrency",
            "type": "integer"
          },
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          },
          "groups": {
            "items": {
              "type": "string"
            },
            "title": "Groups",
            "type": "array"
          },
          "maxFiles": {
            "title": "Maxfiles",
            "type": "integer"
          },
          "selection": {
            "default": "DEFAULT",
            "title": "Selection",
            "type": "string"
          },
          "sourceIds": {
            "items": {
              "type": "string"
            },
            "title": "Sourceids",
            "type": "array"
          }
        },
        "title": "AnalysisBuildRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/knowledge/analysis/build": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Analysis Build Api V1 Knowledge Analysis Build Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/knowledge/analysis/diagnostics": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "sourceId",
            "required": false,
            "schema": {
              "title": "Sourceid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500.0,
              "minimum": 1.0,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0.0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "requestBody": null,
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Analysis Diagnostics Api V1 Knowledge Analysis Diagnostics Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/knowledge/analysis/files": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "sourceId",
            "required": false,
            "schema": {
              "title": "Sourceid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "pathContains",
            "required": false,
            "schema": {
              "title": "Pathcontains",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000.0,
              "minimum": 1.0,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0.0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "requestBody": null,
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Analysis Files Api V1 Knowledge Analysis Files Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/knowledge/analysis/graph/edge/{edge_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "edge_id",
            "required": true,
            "schema": {
              "title": "Edge Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "graphRevision",
            "required": true,
            "schema": {
              "title": "Graphrevision",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sourceId",
            "required": false,
            "schema": {
              "title": "Sourceid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeEvidence",
            "required": false,
            "schema": {
              "default": false,
              "title": "Includeevidence",
              "type": "boolean"
            }
          }
        ],
        "requestBody": null,
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/knowledge/analysis/graph/edges": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "graphRevision",
            "required": true,
            "schema": {
              "title": "Graphrevision",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "title": "Cursor",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 1000,
              "maximum": 5000.0,
              "minimum": 1.0,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sourceId",
            "required": false,
            "schema": {
              "title": "Sourceid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "flowDomain",
            "required": false,
            "schema": {
              "title": "Flowdomain",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "factOrigin",
            "required": false,
            "schema": {
              "title": "Factorigin",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "edgeType",
            "required": false,
            "schema": {
              "title": "Edgetype",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeExternal",
            "required": false,
            "schema": {
              "default": "show",
              "title": "Includeexternal",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeUnresolved",
            "required": false,
            "schema": {
              "default": true,
              "title": "Includeunresolved",
              "type": "boolean"
            }
          }
        ],
        "requestBody": null,
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/knowledge/analysis/graph/manifest": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "sourceId",
            "required": false,
            "schema": {
              "title": "Sourceid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "flowDomain",
            "required": false,
            "schema": {
              "title": "Flowdomain",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "factOrigin",
            "required": false,
            "schema": {
              "title": "Factorigin",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "nodeKind",
            "required": false,
            "schema": {
              "title": "Nodekind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "edgeType",
            "required": false,
            "schema": {
              "title": "Edgetype",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeExternal",
            "required": false,
            "schema": {
              "default": "show",
              "title": "Includeexternal",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeUnresolved",
            "required": false,
            "schema": {
              "default": true,
              "title": "Includeunresolved",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "includeIsolated",
            "required": false,
            "schema": {
              "default": true,
              "title": "Includeisolated",
              "type": "boolean"
            }
          }
        ],
        "requestBody": null,
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/knowledge/analysis/graph/node/{node_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "graphRevision",
            "required": true,
            "schema": {
              "title": "Graphrevision",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sourceId",
            "required": false,
            "schema": {
              "title": "Sourceid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "includeEvidence",
            "required": false,
            "schema": {
              "default": false,
              "title": "Includeevidence",
              "type": "boolean"
            }
          }
        ],
        "requestBody": null,
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        }
      }
    },
...
    "/api/v1/knowledge/analysis/jobs/{job_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": null,
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Analysis Job Api V1 Knowledge Analysis Jobs  Job Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/knowledge/analysis/jobs/{job_id}/stop": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": null,
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Analysis Job Stop Api V1 Knowledge Analysis Jobs  Job Id  Stop Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        }
      }
    },
...
  }
}